# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: ADS WebApp Baseline
# during dev, should point to your local machine
host: localhost:64003
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /auth/refresh:
    x-swagger-router-controller: auth
    post:
      description: Issues a new JWT for user access given a refresh token
      operationId: refresh
      parameters:
      - name: body
        in: body
        description: The refresh token of a successful prior login
        required: true
        schema:
          $ref: "#/definitions/Refresh"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResult"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Refresh:
    properties:
      refresh_token:
        type: string
...
  LoginResult:
    properties:
      access_token:
        type: string
      token_type:
        type: string
      refresh_token:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string